requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
pydantic
